Waiting for real selsia account



See merge request !14

+++ database/migrations/2016_03_02_131832_create_table_used_cars.php
@@ -14,7 +14,7 @@ class CreateTableUsedCars extends Migration {
     {
         Schema::create('used_cars', function (Blueprint $table) {
             $table->increments('id')->unsigned();
-            $table->integer('provider_car_id');
+            $table->bigInteger('provider_car_id');
             $table->string('provider', 20);
             $table->string('marque', 15)->nullable();
             $table->string('model', 100)->nullable();
